+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [0.22.13]
 ### Changed
-- establish x-service-tag-preference on ingress: use the request one if it's more specific, otherwise the default one
+- establish x-service-tag-preference on ingress: use the default one if it's more specific, otherwise the request one
 
 ## [0.22.12]
 ### Changed